test: copy shader assets to cmake binary dir
Yes, a more efficient way might be to just change the working directory
when running tests, but that requires me and everyone else to know to do
that, and to remember to do that. This copy takes no time and very
little space and avoids running into this error which will cause an hour
of furstration to figure out what is wrong and fix it.